Add `Pkg.test()` test, fix `Base.preferences_names` bug
This ensures that things that users want to do within a `Pkg.test()`
invocation work as expected; they can set and load preferences, they can
read preferences that are set in their package before the test run, and
the test does not modify any local files.

This also fixes #31, where `JuliaLocalPreferences.toml` files were
ignored when writing preferences out, but not when reading them in.