Prefer setting preferences only within the currently active project
Previous to this change, we would search up the environment stack
looking for the project that contains the target package, and set the
preference in there.  This was intended to make it convenient to e.g.
set a `Revise` preference while having an application project activated.
It turns out that the more common problem is actually that a package
wants to set a preference of a sub-package (e.g. `MPI` and
`MPIPreferences`) that may not be a top-level dependent of the currently
active project.

Therefore, we change the behavior of `set_preferences!()` to prefer
adding the target package as an `extras` dependency of the
currently-active project, but maintain the old behavior behind a keyword
argument to `set_preferences!()`.